requirements.txt
Django==1.4.2
Mako==0.7.2
MarkupSafe==0.15
distribute==0.6.28
dj-database-url==0.2.1
django-jsonfield==0.8.11
django-mako==0.1.5pre
django-social-auth==0.7.8
httplib2==0.7.6
oauth2==1.5.211
psycopg2==2.4.5
python-openid==2.2.5
wsgiref==0.1.2
